new_backnd/create_meme_video_github/requirements.txt
requests>=2.31.0
google-genai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6